fix(radio): trigger refresh immediately upon checking a radio button
When a radio button is clicked, the change in value should immediately be
notified to the field, possibly triggering a refresh if modeled as such.
Before this fix however, the change in value would only be processed once
focus was lost, e.g. by clicking outside of the radio button or tabbing
to the next element. This would inadvertently delay the refresh.

Fixes BQ-10162