fix: redirect to root after logging out from a specific session
When a session id is present in the URL and the user logs out, the
URL with session id should not be redirected to after having logged
in again as that particular session will have been terminated. Instead,
this commit changes the behavior to redirect to the root instead if an
explicit session id had been used to start the session.

Fixes CSD-3785